+++ lib/commercetools-api/src/Models/Common/BaseResource.php
@@ -91,6 +91,7 @@ use Commercetools\Api\Models\Message\CartDiscountStoreAddedMessage;
 use Commercetools\Api\Models\Message\CartDiscountStoreRemovedMessage;
 use Commercetools\Api\Models\Message\CartDiscountStoresSetMessage;
 use Commercetools\Api\Models\Message\CartFrozenMessage;
+use Commercetools\Api\Models\Message\CartPurchaseOrderNumberSetMessage;
 use Commercetools\Api\Models\Message\CartUnfrozenMessage;
 use Commercetools\Api\Models\Message\CategoryCreatedMessage;
 use Commercetools\Api\Models\Message\CategorySlugChangedMessage;