+++ lib/commercetools-api/src/Models/Order/OrderFromCartDraftBuilder.php
@@ -133,8 +133,9 @@ final class OrderFromCartDraftBuilder implements Builder
     }
 
     /**
-     * <p>User-defined identifier for a purchase Order.</p>
-     * <p>It is typically set by the <a href="ctp:api:type:Buyer">Buyer</a> and can be used with <a href="/quotes-overview">Quotes</a> to track the purchase Order during the <a href="/../api/quotes-overview#intended-workflow">quote and order flow</a>.</p>
+     * <p>User-defined identifier for a purchase order.</p>
+     * <p>It is typically set by the <a href="ctp:api:type:Buyer">Buyer</a> or Merchant to track the purchase order during the <a href="/../api/quotes-overview#intended-workflow">quote and order flow</a>.</p>
+     * <p>If not provided, the <code>purchaseOrderNumber</code> from the referenced <a href="ctp:api:type:Cart">Cart</a> is used.</p>
      *
 
      * @return null|string