+++ lib/commercetools-api/src/Models/Quote/QuoteModel.php
@@ -1101,7 +1101,7 @@ final class QuoteModel extends JsonObjectModel implements Quote
     }
 
     /**
-     * <p>The Purchase Order Number is typically set by the <a href="/quotes-overview#buyer">Buyer</a> on a <a href="ctp:api:type:QuoteRequest">QuoteRequest</a> to
+     * <p>The purchase order number is typically set by the <a href="/quotes-overview#buyer">Buyer</a> on a <a href="ctp:api:type:QuoteRequest">QuoteRequest</a> to
      * track the purchase order during the <a href="/../api/quotes-overview#intended-workflow">quote and order flow</a>.</p>
      *
      *