+++ lib/commercetools-api/src/Models/QuoteRequest/QuoteRequestBuilder.php
@@ -614,7 +614,7 @@ final class QuoteRequestBuilder implements Builder
 
     /**
      * <p>Identifier for a purchase order, usually in a B2B context.
-     * The Purchase Order Number is typically entered by the <a href="/quotes-overview#buyer">Buyer</a>.</p>
+     * The purchase order number is typically entered by the <a href="/quotes-overview#buyer">Buyer</a>.</p>
      *
 
      * @return null|string