+++ lib/commercetools-api/src/Models/StagedQuote/StagedQuoteBuilder.php
@@ -316,7 +316,7 @@ final class StagedQuoteBuilder implements Builder
     }
 
     /**
-     * <p>The Purchase Order Number is typically set by the <a href="/quotes-overview#buyer">Buyer</a> on a <a href="ctp:api:type:QuoteRequest">QuoteRequest</a> to
+     * <p>The purchase order number is typically set by the <a href="/quotes-overview#buyer">Buyer</a> on a <a href="ctp:api:type:QuoteRequest">QuoteRequest</a> to
      * track the purchase order during the <a href="/../api/quotes-overview#intended-workflow">quote and order flow</a>.</p>
      *